fix: a compound knob expression cast its result to Float64
The last -- and worst -- instance of the bug class this sweep has been
chasing. `_eval_knob_node` for a KnobCall ended with

    return Float64(f(vals...))

so `@knob_expr -(bus + trim)` evaluated the arithmetic correctly, in
whatever type the knobs held, and then threw the type away. Worst because
the VALUE stayed right: a derivative seeded on a knob would come back
exactly zero with nothing to indicate anything had happened. A bare
`@knob_expr bus` was unaffected, which is the kind of asymmetry nobody
finds by reading.

This matters more than the element-level pins it follows. A knob is a
machine control -- one supply feeding a whole magnet family -- so a
derivative with respect to a knob is a family's response to its power
supply, which is what orbit correction and matching actually want. It is
also the thing a per-placement override cannot express, since an override
detaches a magnet from its family by design.

Measured with real ForwardDiff 1.4.3, one knob feeding two quadrupoles:

    d(x_out)/d(bus), both magnets moving  -0.0007299630316598207
    central difference                    -0.0007299630318063043

and it agrees with the existing symbolic route: knob_derivative for
bus*bus + trim gives bus + bus, evaluating to 6.0 at bus = 3.

Note the knob type declaration is load-bearing and stays that way: knobs
default to Float64, so `@knob bus` still refuses a dual and `@knob
bus::Real` is what an AD seed needs. That is a feature -- a machine
setting should be typed -- and it is now documented by the test rather
than discovered.

The regression guard uses BigFloat rather than a dual, so it needs no AD
dependency; both are simply "not Float64", and the cast would fail either
identically. It also pins that a Float64 knob still narrows, and that a
Float64-valued knob still yields a Float64 element.

Behavioural fingerprint byte-identical, same sha256. Suite green, Knob
control 56/56.

+++ src/knobs/Knobs.jl
@@ -563,7 +563,13 @@ function _eval_knob_node(n::KnobCall, active::Set{Symbol})
             "$(repr(v))::$(typeof(v)); only Real-typed knobs can appear inside " *
             "arithmetic expressions"))
     end
-    return Float64(f(vals...))
+    # NOT `Float64(...)`: a knob may hold a dual number or a truncated power
+    # series, and casting here would evaluate the expression correctly and then
+    # throw the derivative away -- silently, since the value would be right.
+    # A knob is a machine control, so differentiating with respect to one is the
+    # most useful derivative there is; `bus` feeding a whole magnet family gives
+    # the family's response in a single pass.
+    return f(vals...)
 end
 
 # ---------------------------------------------------------------------------
